+++ src/Tests/Unit/Public/Reset-ModelContext.Tests.ps1
@@ -15,7 +15,8 @@ InModuleScope 'pwshBedrock' {
             $_.ModelId -ne 'amazon.nova-canvas-v1:0' -and
             $_.ModelId -ne 'amazon.nova-reel-v1:0' -and
             $_.ModelId -ne 'cohere.command-text-v14' -and
-            $_.ModelId -ne 'cohere.command-light-text-v14'
+            $_.ModelId -ne 'cohere.command-light-text-v14' -and
+            $_.ModelId -ne 'luma.ray-v2:0'
         }).ModelID
     #-------------------------------------------------------------------------
     $WarningPreference = "SilentlyContinue"
@@ -375,6 +376,20 @@ User: "Hi there!
                             }
                         )
                     }
+                    # [PSCustomObject]@{
+                    #     ModelID = 'luma.ray-v2:0'
+                    #     Context = @(
+                    #         [PSCustomObject]@{
+                    #             role    = 'user'
+                    #             content = @(
+                    #                 [PSCustomObject]@{
+                    #                     type = 'text'
+                    #                     text = 'Ray v2 context'
+                    #                 }
+                    #             )
+                    #         }
+                    #     )
+                    # }
                     [PSCustomObject]@{
                         ModelID = 'meta.llama3-8b-instruct-v1:0'
                         Context = @(