+++ src/pwshBedrock/Public/Get-ModelInfo.ps1
@@ -69,6 +69,7 @@ function Get-ModelInfo {
             'cohere.command-light-text-v14',
             'cohere.command-r-v1:0',
             'cohere.command-r-plus-v1:0',
+            'luma.ray-v2:0',
             'meta.llama3-70b-instruct-v1:0',
             'meta.llama3-8b-instruct-v1:0',
             'meta.llama3-1-8b-instruct-v1:0',
@@ -117,6 +118,7 @@ function Get-ModelInfo {
     $modelInfo += $script:amazonModelInfo
     $modelInfo += $script:ai21ModelInfo
     $modelInfo += $script:cohereModelInfo
+    $modelInfo += $script:lumaModelInfo
     $modelInfo += $script:metaModelInfo
     $modelInfo += $script:mistralAIModelInfo
     $modelInfo += $script:stabilityAIModelInfo