+++ src/pwshBedrock/Public/Reset-ModelContext.ps1
@@ -66,6 +66,7 @@ function Reset-ModelContext {
             # 'cohere.command-light-text-v14',
             'cohere.command-r-v1:0',
             'cohere.command-r-plus-v1:0',
+            'luma.ray-v2:0',
             'meta.llama3-70b-instruct-v1:0',
             'meta.llama3-8b-instruct-v1:0',
             'meta.llama3-1-8b-instruct-v1:0',
@@ -160,7 +161,8 @@ function Reset-ModelContext {
                             $_.ModelId -ne 'amazon.nova-canvas-v1:0' -and
                             $_.ModelId -ne 'amazon.nova-reel-v1:0' -and
                             $_.ModelId -ne 'cohere.command-text-v14' -and
-                            $_.ModelId -ne 'cohere.command-light-text-v14'
+                            $_.ModelId -ne 'cohere.command-light-text-v14' -and
+                            $_.ModelId -ne 'luma.ray-v2:0'
                         }).ModelID
                     foreach ($model in $allModelIDs) {
                         $context = $Global:pwshBedrockModelContext | Where-Object { $_.ModelID -eq $model }