+++ docs/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.2.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.55.0]
+
+- Module changes:
+    - Luma
+        - Added new function `Invoke-LumaAIModel` to support video model: `luma.ray-v2:0`
+    - Manifest updates:
+        - Require more recent version of `AWS.Tools.BedrockRuntime` to take advantage of new commands such as `Start-BDRRAsyncInvoke`
+        - Require `AWS.Tools.S3` to support async video file generation download from S3.
+- Build changes:
+    - Updated bootstrap file to latest versions of modules
+
 ## [0.52.0] - **BREAKING CHANGES**
 
 - Module changes: